Feat/aws wifi pubsub  
(Renesas members can see the products/common/rtos/amazon-freertos/lts/iot-reference-rl78!11 in the development repo. for details.)

+++ Projects/rl78g23-fpb/modules/r_wifi_da16xxx/src/r_wifi_da16xxx_os_wrap.c
@@ -343,9 +343,9 @@ void tick_count_stop(void)
  * Return Value : TICK_NOT_EXPIERD
  *                TICK_EXPIERD
  *********************************************************************************************************************/
-uint32_t tick_count_check(void)
+uint8_t tick_count_check(void)
 {
-    uint32_t ret = TICK_NOT_EXPIRED;
+    uint8_t ret = TICK_NOT_EXPIRED;
     OS_TICK tick_tmp;
 
     if (1 != s_exec_tick_count_flg)
